move serveCache code to the plugin folder
